# pkg/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Ride-Hailing API
  description: |
    Comprehensive ride-hailing platform API supporting riders, drivers, and administrators.

    ## Authentication
    All protected endpoints require a Bearer JWT token in the Authorization header.
    Obtain tokens via the `/api/v1/auth/login` or `/api/v1/auth/register` endpoints.

    ## Rate Limiting
    All endpoints are rate limited. Check `X-RateLimit-Limit`, `X-RateLimit-Remaining`,
    and `X-RateLimit-Reset` response headers.

    ## Idempotency
    For POST/PUT/PATCH requests, include an `Idempotency-Key` header to prevent duplicate processing.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Auth Service
  - url: http://localhost:8081
    description: Rides Service
  - url: http://localhost:8083
    description: Geo Service
  - url: http://localhost:8084
    description: Payments Service
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    DriverLocation:
      type: object
      properties:
        driver_id:
          type: string
          format: uuid
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        h3_cell:
          type: string
        heading:
          type: number
          format: double
        speed:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
...
    GeocodingResult:
      type: object
      properties:
        place_id:
          type: string
        formatted_address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        types:
          type: array
          items:
            type: string
        h3_cell:
          type: string
    AutocompleteResult:
      type: object
      properties:
        place_id:
          type: string
        description:
          type: string
        main_text:
          type: string
        secondary_text:
          type: string
    SurgeInfo:
      type: object
      properties:
        h3_cell:
          type: string
        surge_multiplier:
          type: number
          format: double
        demand_count:
          type: integer
        supply_count:
          type: integer
        updated_at:
          type: string
          format: date-time
  parameters:
...
    LatitudeParam:
      name: latitude
      in: query
      required: true
      schema:
        type: number
        format: double
    LongitudeParam:
      name: longitude
      in: query
      required: true
      schema:
        type: number
        format: double
paths:
...
  /api/v1/geo/location:
    post:
      tags: [Geo]
      summary: Update driver location
      description: Drivers only. Updates location with H3 spatial indexing.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude:
                  type: number
                  format: double
                longitude:
                  type: number
                  format: double
                heading:
                  type: number
                  format: double
                  description: Direction 0-360 degrees
                speed:
                  type: number
                  format: double
                  description: Speed in km/h
      responses:
        '200':
          description: Location updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                          h3_cell:
                            type: string
  /api/v1/geo/drivers/{id}/location:
    get:
      tags: [Geo]
      summary: Get driver location
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Driver location
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DriverLocation'
  /api/v1/geo/drivers/nearby:
    get:
      tags: [Geo]
      summary: Find nearby available drivers
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LatitudeParam'
        - $ref: '#/components/parameters/LongitudeParam'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: Nearby drivers
  /api/v1/geo/distance:
    post:
      tags: [Geo]
      summary: Calculate distance and ETA between two points
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_latitude, from_longitude, to_latitude, to_longitude]
              properties:
                from_latitude:
                  type: number
                  format: double
                from_longitude:
                  type: number
                  format: double
                to_latitude:
                  type: number
                  format: double
                to_longitude:
                  type: number
                  format: double
      responses:
        '200':
          description: Distance and ETA
  /api/v1/geo/h3/cell:
    get:
      tags: [Geo]
      summary: Get H3 cell indexes at all resolutions for a location
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LatitudeParam'
        - $ref: '#/components/parameters/LongitudeParam'
      responses:
        '200':
          description: H3 cell indexes
  /api/v1/geo/h3/surge:
    get:
      tags: [Geo]
      summary: Get surge pricing info for a location (H3-based)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LatitudeParam'
        - $ref: '#/components/parameters/LongitudeParam'
      responses:
        '200':
          description: Surge pricing info
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SurgeInfo'
  /api/v1/geo/h3/demand-heatmap:
    get:
      tags: [Geo]
      summary: Get demand heatmap data around a location
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LatitudeParam'
        - $ref: '#/components/parameters/LongitudeParam'
      responses:
        '200':
          description: Demand heatmap
  /api/v1/geo/geocode:
    get:
      tags: [Geo]
      summary: Forward geocode an address to coordinates
      security:
        - BearerAuth: []
      parameters:
        - name: address
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Geocoding results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          results:
                            type: array
                            items:
                              $ref: '#/components/schemas/GeocodingResult'
  /api/v1/geo/geocode/reverse:
    get:
      tags: [Geo]
      summary: Reverse geocode coordinates to an address
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LatitudeParam'
        - $ref: '#/components/parameters/LongitudeParam'
      responses:
        '200':
          description: Reverse geocoding results
  /api/v1/geo/geocode/autocomplete:
    get:
      tags: [Geo]
      summary: Place search autocomplete
      description: Returns place suggestions as the user types. Optionally biased toward a location.
      security:
        - BearerAuth: []
      parameters:
        - name: input
          in: query
          required: true
          schema:
            type: string
        - name: latitude
          in: query
          schema:
            type: number
            format: double
        - name: longitude
          in: query
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Autocomplete predictions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          predictions:
                            type: array
                            items:
                              $ref: '#/components/schemas/AutocompleteResult'
  /api/v1/geo/geocode/place:
    get:
      tags: [Geo]
      summary: Get place details by place ID
      security:
        - BearerAuth: []
      parameters:
        - name: place_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Place details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/GeocodingResult'

  # ==================== PROMOS ====================
...